requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.37.0
accelerate>=0.27.0
Pillow>=9.0.0
numpy>=1.24.0
tqdm>=4.65.0
datasets>=2.18.0
einops>=0.7.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# LLaVA (install separately from source):
#   git clone https://github.com/haotian-liu/LLaVA && pip install -e LLaVA
